Refactored edit-item-quantity, item-basic-stats, item-summary-info, sidebar, sidebar-nav-item, and statistics-tooltip components
