+++ src/Components/item-summary-info/item-summary-info.styles.scss
@@ -1,0 +1,26 @@
+@import "../../Styles/base/_settings.scss";
+
+.list-header {
+  background-color: $gray-color;
+  color: $primary-color;
+  font-weight: 700;
+  letter-spacing: 0.15rem;
+  text-transform: uppercase;
+  font-size: smaller;
+}
+
+.list-row {
+  display: flex;
+  justify-content: space-between;
+  font-size: small;
+
+  span:first-child {
+    font-weight: 500;
+  }
+}
+
+.stat-value {
+  font-family: "Work Sans";
+  letter-spacing: 0.07em;
+  font-size: 12px;
+}
