+++ src/Components/sidebar/sidebar.component.js
@@ -1,22 +1,16 @@
 import React from "react";
+//  COMPONENT IMPORTS
 import { Nav, NavItem } from "react-bootstrap";
 import { IndexLinkContainer } from "react-router-bootstrap";
-import NavLink from "./SidebarNavItem";
+import NavLink from "../sidebar-nav-item/sidebar-nav-item.component";
+import "./sidebar.styles.scss";
 
-const Sidebar = () => {
+const SidebarComponent = () => {
   return (
     <Nav className="flex-column sidebar">
       <input
         type="text"
         placeholder="Search"
-        style={{
-          padding: "20px",
-          height: "10px",
-          width: "80%",
-          margin: "20px 10px 10px 10px",
-          borderRadius: "1px",
-          border: "1px solid white"
-        }}
       />
       <IndexLinkContainer to="/" activeClassName="activeLink">
         <NavItem>{"Dashboard"}</NavItem>
@@ -31,4 +25,4 @@ const Sidebar = () => {
   );
 };
 
-export default Sidebar;
+export default SidebarComponent;
