+++ src/Routers/Router.js
@@ -8,7 +8,7 @@ import Cobas8100 from "../Pages/Cobas8100";
 import ManualUsedTransaction from "../Pages/ManualUsedTransaction";
 import Error404 from "../Pages/404";
 import ItemDetails from "../Pages/ItemDetails";
-import Sidebar from "../Components/Sidebar";
+import SidebarComponent from "../components/sidebar/sidebar.component";
 import { ArchitectContext } from "../Contexts/ArchitectContext";
 import { ABLContext } from "../Contexts/ABLContext";
 import Login from "../Login";
@@ -21,7 +21,7 @@ const Router = () => {
   return (
     <BrowserRouter>
       <div className="app-container">
-        <Sidebar />
+        <SidebarComponent />
         <section className="content-container">
           <Switch>
             <Route exact path="/signup" component={SignUp} />
