+++ src/Styles/styles.scss
@@ -1,6 +1,5 @@
 @import url("https://fonts.googleapis.com/css?family=Poppins:400,500,700,800,900&display=swap");
 @import "./base/settings";
-@import "./Components/sidebar.scss";
 @import "./Components/itemdetails.scss";
 @import "./Components/instruments.scss";
